server renamed to api and docs added
