+++ src/maker/handlers.rs
@@ -13,7 +13,7 @@ use crate::protocol::{
 };
 
 use crate::{
-    maker::server::ExpectedMessage,
+    maker::api::ExpectedMessage,
     protocol::{
         contract::{
             calculate_coinswap_fee, create_receivers_contract_tx, find_funding_output_index,
@@ -31,8 +31,8 @@ use crate::{
 };
 
 use super::{
+    api::{ConnectionState, Maker, MakerBehavior},
     error::MakerError,
-    server::{ConnectionState, Maker, MakerBehavior},
 };
 
 /// The Global Handle Message function. Takes in a [Arc<Maker>] and handle messages
